# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Lançamento e Prestação de Contas",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/suppliers": {
      "get": {
        "tags": [
          "fornecedores"
        ],
        "summary": "Buscar",
        "description": "Busca por similaridade de nome (tolerante a acentos e erros de digitação) ou CNPJ.",
        "operationId": "buscar_api_v1_suppliers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Nome ou início do CNPJ",
              "title": "Q"
            },
            "description": "Nome ou início do CNPJ"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierOut"
                  },
                  "title": "Response Buscar Api V1 Suppliers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "fornecedores"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_suppliers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suppliers/{supplier_id}": {
      "get": {
        "tags": [
          "fornecedores"
        ],
        "summary": "Obter",
        "operationId": "obter_api_v1_suppliers__supplier_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Supplier Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "fornecedores"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_suppliers__supplier_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Supplier Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fornecedores"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_suppliers__supplier_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Supplier Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/lancamentos": {
      "get": {
        "tags": [
          "lançamentos"
        ],
        "summary": "Listar",
        "description": "Lista ordenada por vencimento (mais próximo primeiro; sem vencimento no fim).",
        "operationId": "listar_api_v1_lancamentos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusEfetivo"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Status efetivo (repita para vários)",
              "title": "Status"
            },
            "description": "Status efetivo (repita para vários)"
          },
          {
            "name": "supplier_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier Id"
            }
          },
          {
            "name": "categoria_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoria Id"
            }
          },
          {
            "name": "projeto_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projeto Id"
            }
          },
          {
            "name": "centro_custo_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Centro Custo Id"
            }
          },
          {
            "name": "vencimento_de",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vencimento De"
            }
          },
          {
            "name": "vencimento_ate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vencimento Ate"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LancamentoPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "lançamentos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_lancamentos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LancamentoIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LancamentoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lancamentos/{lanc_id}": {
      "get": {
        "tags": [
          "lançamentos"
        ],
        "summary": "Obter",
        "operationId": "obter_api_v1_lancamentos__lanc_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lanc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lanc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LancamentoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "lançamentos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_lancamentos__lanc_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lanc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lanc Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Versão esperada, ex.: \"3\"",
              "title": "If-Match"
            },
            "description": "Versão esperada, ex.: \"3\""
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LancamentoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LancamentoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "lançamentos"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_lancamentos__lanc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lanc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lanc Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LancamentoIn": {
        "properties": {
          "valor": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Valor",
            "examples": [
              "245.90"
            ]
          },
          "data_emissao": {
            "type": "string",
            "format": "date",
            "title": "Data Emissao"
          },
          "data_pagamento_prevista": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Pagamento Prevista"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "projeto_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projeto Id"
          },
          "centro_custo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Centro Custo Id"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "forma_pagamento": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "boleto",
                  "pix",
                  "cartao",
                  "dinheiro",
                  "transferencia",
                  "outro"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Forma Pagamento"
          },
          "linha_digitavel": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Linha Digitavel"
          }
        },
        "type": "object",
        "required": [
          "valor",
          "data_emissao"
        ],
        "title": "LancamentoIn"
      },
      "LancamentoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "valor": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Valor"
          },
          "data_emissao": {
            "type": "string",
            "format": "date",
            "title": "Data Emissao"
          },
          "data_pagamento_prevista": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Pagamento Prevista"
          },
          "data_pagamento_efetiva": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Pagamento Efetiva"
          },
          "supplier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupplierRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "categoria": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Ref"
              },
              {
                "type": "null"
              }
            ]
          },
          "projeto": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Ref"
              },
              {
                "type": "null"
              }
            ]
          },
          "centro_custo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CentroCustoRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "usuario": {
            "$ref": "#/components/schemas/Ref"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "forma_pagamento": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "boleto",
                  "pix",
                  "cartao",
                  "dinheiro",
                  "transferencia",
                  "outro"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Forma Pagamento"
          },
          "linha_digitavel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linha Digitavel"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_efetivo": {
            "$ref": "#/components/schemas/StatusEfetivo"
          },
          "dias_para_vencimento": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Para Vencimento"
          },
          "status_aprovacao": {
            "type": "string",
            "title": "Status Aprovacao"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "valor",
          "data_emissao",
          "data_pagamento_prevista",
          "data_pagamento_efetiva",
          "supplier",
          "categoria",
          "projeto",
          "centro_custo",
          "usuario",
          "descricao",
          "forma_pagamento",
          "linha_digitavel",
          "status",
          "status_efetivo",
          "dias_para_vencimento",
          "status_aprovacao",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "LancamentoOut"
      },
      "LancamentoPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LancamentoOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "LancamentoPage"
      },
...
      "StatusEfetivo": {
        "type": "string",
        "enum": [
          "pendente",
          "reagendado",
          "vence_hoje",
          "atrasado",
          "pago",
          "rejeitado"
        ],
        "title": "StatusEfetivo"
      },
      "SupplierIn": {
        "properties": {
          "nome_fantasia": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Nome Fantasia"
          },
          "razao_social": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Razao Social"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj",
            "examples": [
              "11.222.333/0001-81",
              "12.ABC.345/01DE-35"
            ]
          }
        },
        "type": "object",
        "required": [
          "nome_fantasia"
        ],
        "title": "SupplierIn"
      },
      "SupplierOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome_fantasia": {
            "type": "string",
            "title": "Nome Fantasia"
          },
          "razao_social": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Razao Social"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "cnpj_formatado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj Formatado",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome_fantasia",
          "razao_social",
          "cnpj",
          "created_at",
          "cnpj_formatado"
        ],
        "title": "SupplierOut"
      },
      "SupplierRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome_fantasia": {
            "type": "string",
            "title": "Nome Fantasia"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome_fantasia",
          "cnpj"
        ],
        "title": "SupplierRef"
      },
      "SupplierUpdate": {
        "properties": {
          "nome_fantasia": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome Fantasia"
          },
          "razao_social": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Razao Social"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          }
        },
        "type": "object",
        "title": "SupplierUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}